Refactor and clean all old scaled charts usages
